requirements.txt
fastapi
uvicorn
pypdf
python-docx
docx2txt
chromadb
httpx
python-multipart
sentence-transformers
langchain
